Minor modifications, tests added and organized.
New tests were aded and existing tests for running and installing new
versions are also modified to take in account if there is a version
specified in the .nvmerc config file.
Minor modification to .travis.yml so that the language is c, which
defaults to ruby, instead of bash.
Corrected lower- to uppercase for $NVME_DIR in bash_completion.